add stylecom parser, add caching, little improve logic

+++ search_service/parsers/pipline.py
@@ -1,29 +1,51 @@
 import os
 from datetime import datetime
-from functools import cache
 
 from search_service.parsers.laptop.forlaptop import ForLaptopKievParser
+from search_service.parsers.laptop.fornb import ForNBParser
 from search_service.parsers.phone.all_spares import AllSparesParser
 from search_service.parsers.phone.motorolka import MotorolkaParser
+from search_service.parsers.phone.stylecom import StylecomParser
 
 
-@cache
+def delete_old_files():
+    current_time = datetime.now()
+    for filename in os.listdir("."):
+        if filename.endswith(".csv"):
+            file_time = os.path.getmtime(filename)
+            file_datetime = datetime.fromtimestamp(file_time)
+            if (current_time - file_datetime).seconds > 60 * 60 * 16:  # 16 hours
+                os.remove(filename)
+
+
+def check_file_exists(filename: str) -> bool:
+    delete_old_files()
+    return os.path.exists(filename)
+
+
+def make_filename(query: str) -> str:
+    query = query.replace(" ", "_")
+    return f"{query}.csv"
+
+
 def start_pipline(query: str):
-    time_ = datetime.today().strftime("%Y-%m-%d_%H-%M")
+    finalname = make_filename(query)
+    if check_file_exists(finalname):
+        print(f"File {finalname} already exists. Returning existing file.")
+        return finalname
+
     parsers = [
         ForLaptopKievParser(query),
         MotorolkaParser(query),
         AllSparesParser(query),
+        ForNBParser(query),
+        StylecomParser(query),
     ]
 
     for parser in parsers:
         print(f"Start parser: {parser.filename}")
         parser.parse()
         print(f"Finished parser: {parser.filename}")
-
-    query = query.replace(" ", "_")
-
-    finalname = f"{query}_{time_}.csv"
 
     if os.path.exists(finalname):
         os.remove(finalname)
@@ -34,8 +56,5 @@ def start_pipline(query: str):
                 with open(parser.filename, "r") as file:
                     lines = file.readlines()
                     f.writelines(lines[1:])  # Пропускаємо заголовок
-
-    # with open(finalname, "r") as f:
-    #     print(f.read())
 
     return finalname
